# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MoobilPay API",
        "description": "API complète de paiement en ligne avec Stripe et PayPal. Permet de gérer les paiements, méthodes de paiement et webhooks de manière sécurisée.",
        "contact": {
            "name": "Support MoobilPay",
            "email": "tchindavaldoblair@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://api.moobilpay.com",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/payments/{uuid}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Détails d'un paiement",
                "operationId": "441d476ac4eed40e33d1dbdf17659bf8",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du paiement"
                    }
                }
            }
        },
        "/api/payments/{uuid}/confirm": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Confirmer un paiement",
                "operationId": "412632014ac31fb496fc6254d8fab1e1",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paiement confirmé"
                    }
                }
            }
        },
        "/api/payments/{uuid}/refund": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Rembourser un paiement",
                "operationId": "1fb02f6689275ea4333d3e791d768240",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Remboursement effectué"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification via Laravel Sanctum. Utilisez votre token d'API.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}